requirements.txt
pyyaml
requests
pandas
python-dotenv
scikit-learn
joblib
numpy
xgboost
lightgbm